+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Add sale order and client partner fields on production order.
 - COST SHEET : calculation date
 - INVOICE : filter on product in InvoiceLine on supplier catalog for supplier invoice
 - Message : Email - 'From' e-mail adress contact Name.
+- COST SHEET : calculation rules
 
 ## Improvements
 - FINANCIAL ACCOUNT : New boolean 'Analytic Distribution Authorized' added to make Analytic fields required on Invoice Line.